Write a function that takes a pointer to an int as parameter and updates the value it points to to 98

+++ 0x05-pointers_arrays_strings/main.h
@@ -1,6 +1,6 @@
 void reset_to_98(int *n);
         void swap_int(int *a, int *b);
-	size_t _strlen(char *s);
+	int _strlen(char *s);
 	void _puts(char *str);
 	int _putchar(char c);
 	void print_rev(char *s);
@@ -9,4 +9,3 @@ void reset_to_98(int *n);
 	void print_arrays(int *a, int n);
 	char *_strcpy(char *dest, char *src);
 	void rev_string(char *s);
-
